Abstract load_data to separate file
